+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+ - Fixed import of shaded guava class via Avro.
+ 
 ## [2.1.0] - 2017-07-24
 ### Added
  - Repo/license badges in README.